+++ runtests.sh
@@ -1,0 +1,2 @@
+python -m doctest vtt_convert.py
+python -m pytest