+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Chagned
+- Each sentence is now capitalized
+
 ## [0.1.1] - 2019-10-06
 
 ### Added