+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Chagned
 - Each sentence is now capitalized
+- Fix typo `vvt_convert` to `vtt_convert` (!incompatible)
 
 ## [0.1.1] - 2019-10-06