+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2019-10-06
+
 ### Added
 - Testing
 
@@ -26,7 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Project starts
 
-[Unreleased]: https://github.com/hiiwave/netflix-vocab-builder/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/hiiwave/netflix-vocab-builder/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/hiiwave/netflix-vocab-builder/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/hiiwave/netflix-vocab-builder/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/hiiwave/netflix-vocab-builder/releases/tag/v0.1.0
 [i1]: https://github.com/hiiwave/netflix-vocab-builder/pull/1